fix: redo theme change slice, make error button component separately, correct styles
