+++ search-app/src/components/errorBoundary/styles.module.css
@@ -1,0 +1,5 @@
+.triggerErrorButton {
+  position: absolute;
+  left: 10px;
+  top: 10px;
+}
