--- search-app/src/components/footer/styles.css
@@ -1,4 +1,0 @@
-.footer {
-  width: 100%;
-  min-height: 50px;
-}
