+++ search-app/src/components/loader/styles.module.css
@@ -1,4 +1,4 @@
-.loader-container {
+.loaderContainer {
   width: 100%;
   flex-grow: 1;
   display: flex;
