+++ search-app/src/components/main/main.tsx
@@ -1,9 +1,10 @@
+import styles from "./styles.module.css";
 import { MainContent } from "./mainContent";
 import { ErrorBoundary } from "../errorBoundary/errorBoundary";
 import { ReactElement } from "react";
 
 export const Main = (): ReactElement => (
-  <main className="main" data-testid="main">
+  <main className={styles.main} data-testid="main">
     <ErrorBoundary>
       <MainContent />
     </ErrorBoundary>
