--- search-app/src/components/main/resultsSection/flyout/styles.css
@@ -1,26 +1,0 @@
-.flyout {
-  position: fixed;
-  bottom: 200px;
-  width: 300px;
-
-  border-radius: 8px;
-  padding: 10px;
-  z-index: 1;
-}
-
-.dark .flyout {
-  background-color: var(--dark);
-  border: 1px solid var(--main-light);
-}
-
-.light .flyout {
-  background-color: var(--light);
-  border: 1px solid var(--main-dark);
-}
-
-.flyout__buttons {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-  padding: 10px 0;
-}
