--- search-app/src/components/main/resultsSection/pagination/styles.css
@@ -1,69 +1,0 @@
-.pagination {
-  display: flex;
-  gap: 10px;
-}
-
-.pagination-numbers {
-  display: flex;
-  gap: 10px;
-}
-
-.pagination-link {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 5px;
-  border-radius: 5px;
-  font-size: 1em;
-  font-weight: 400;
-  font-family: inherit;
-  cursor: pointer;
-  transition: background-color 0.4s ease;
-}
-
-.pagination-link_disabled {
-  pointer-events: none;
-}
-
-.dark .pagination-link {
-  color: var(--main-light);
-  border: 1px solid var(--main-light);
-}
-
-.dark .pagination-link_active {
-  background-color: var(--main-dark-buttons);
-  color: #000000;
-}
-
-.dark .pagination-link_disabled {
-  color: var(--transparent-dark-disabled);
-  border: 1px solid var(--transparent-dark-disabled);
-}
-
-.light .pagination-link {
-  color: var(--main-dark);
-  border: 1px solid var(--main-dark);
-  background-color: var(--transparent-light);
-}
-
-.light .pagination-link_active {
-  background-color: var(--main-light-buttons);
-  color: #000000;
-}
-
-.light .pagination-link_disabled {
-  color: var(--transparent-light-disabled);
-  border: 1px solid var(--transparent-light-disabled);
-}
-
-@media (hover: hover) {
-  .dark .pagination-link:hover {
-    background-color: var(--main-dark-buttons);
-    color: #000000;
-  }
-
-  .light .pagination-link:hover {
-    background-color: var(--main-light-buttons);
-    color: #000000;
-  }
-}
