+++ search-app/src/components/main/resultsSection/peopleList.tsx
@@ -1,18 +1,8 @@
-import { useState } from "react";
 import { Person } from "./person";
 import { PeopleResponse } from "../../../types/types";
 import { Flyout } from "./flyout/flyout";
 
 export const PeopleList = ({ people }: { people: PeopleResponse["results"] }) => {
-  const [isError, setIsError] = useState(false);
-  const triggerError = () => {
-    setIsError(true);
-  };
-
-  if (isError) {
-    throw new Error("An error has occurred");
-  }
-
   return (
     <>
       {people && people.length > 0 ? (
@@ -25,9 +15,6 @@ export const PeopleList = ({ people }: { people: PeopleResponse["results"] }) =>
         <p>The list is empty</p>
       )}
       <Flyout />
-      <button className="trigger-error-button" onClick={triggerError}>
-        Trigger Error
-      </button>
     </>
   );
 };
