--- search-app/src/components/main/resultsSection/styles.css
@@ -1,86 +1,0 @@
-.results-section {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
-  margin: 0 20px;
-  flex-grow: 1;
-}
-
-.results {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  list-style: none;
-  margin: 10px auto;
-  gap: 20px;
-}
-
-.results__item {
-  text-align: start;
-  display: flex;
-  flex-direction: column;
-  border-radius: 10px;
-  box-shadow: 4px 4px 5px var(--border-shadow);
-  padding: 10px;
-  width: 260px;
-  height: 115px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.dark .results__item {
-  background-color: var(--transparent-dark);
-  border: 1px solid var(--main-light);
-}
-
-.light .results__item {
-  background-color: var(--transparent-light);
-  border: 1px solid var(--main-dark);
-}
-
-.name__text,
-.params__text {
-  font-weight: 500;
-}
-
-.name__item {
-  font-weight: 400;
-}
-
-.dark .name__text {
-  color: var(--accent-blue-light);
-}
-
-.dark .params__item {
-  color: var(--text-light);
-  font-weight: 300;
-}
-
-.light .name__text {
-  color: var(--accent-blue-dark);
-}
-
-.light .params__item {
-  color: var(--text-dark);
-  font-weight: 400;
-}
-
-.params__text {
-  display: flex;
-  flex-direction: column;
-}
-
-@media (hover: hover) {
-  .dark .results__item:hover {
-    background-color: var(--transparent-dark-hover);
-    box-shadow: 4px 4px 20px var(--border-shadow);
-  }
-
-  .light .results__item:hover {
-    background-color: var(--transparent-light-hover);
-    box-shadow: 4px 4px 20px var(--border-shadow);
-  }
-}
