--- search-app/src/components/main/searchSection/styles.css
@@ -1,16 +1,0 @@
-.search-form {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-}
-
-.search-input {
-  width: 500px;
-  height: 30px;
-}
-
-.search-button {
-  width: 100px;
-  height: 30px;
-}
