--- search-app/src/components/main/styles.css
@@ -1,42 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  flex-grow: 1;
-}
-
-.search-section::after,
-.footer::before {
-  content: "";
-  display: block;
-  margin: 10px 0;
-}
-
-.dark .search-section::after,
-.dark .footer::before {
-  border-bottom: 1px solid var(--main-light);
-}
-
-.light .search-section::after,
-.light .footer::before {
-  border-bottom: 1px solid var(--main-dark);
-}
-
-.trigger-error-button {
-  position: absolute;
-  left: 10px;
-  top: 10px;
-}
-
-.result-section-with-detail {
-  display: flex;
-  gap: 10px;
-  flex-grow: 1;
-}
-
-.details {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  margin: 5px;
-}
