+++ search-app/src/components/main/styles.module.css
@@ -1,0 +1,6 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  flex-grow: 1;
+}
