--- search-app/src/components/main/toggleTheme/styles.css
@@ -1,5 +1,0 @@
-.toggle-theme_button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-}
