+++ search-app/src/pages/[slug]/Details.tsx
