+++ search-app/src/pages/styles.css
@@ -6,64 +6,59 @@
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 
-  --main-dark: #1d009e;
-  --dark: #090033;
-  --transparent-dark: rgb(9, 0, 51, 0.7);
-  --transparent-dark-hover: rgba(119, 94, 229, 0.2);
-  --transparent-dark-disabled: rgba(119, 94, 229, 0.2);
-  --text-light: #dedbff;
-  --accent-blue-light: #6e75ff;
-  --links-dark-hover: #a0a5ff;
-  --main-dark-buttons: #7378d8;
-
-  --main-light: #ccc1ff;
-  --light: #b4a6f3;
-  --transparent-light: rgba(180, 166, 243, 0.7);
-  --transparent-light-hover: rgba(233, 233, 233, 0.5);
-  --transparent-light-disabled: rgba(233, 233, 233, 0.5);
-  --text-dark: #090033;
-  --accent-blue-dark: #0009b1;
-  --links-dark-hover: #4d54da;
-  --main-light-buttons: #7277ff;
-
-  --main-buttons: #646cff;
-  --border-shadow: rgba(255, 255, 255, 0.5);
-  --links: #646cff;
-  --main-buttons-hover: #a0a5ff;
-  --accent-red: #ff0000;
+  --main-color: #646cff;
+  --main-color-hover: #a0a5ff;
+  --accent-color: #ff0000;
+  --border-shadow-color: rgba(255, 255, 255, 0.5);
   --loader-mask: conic-gradient(#0000 10%, #000), linear-gradient(#000 0 0) content-box;
 }
 
-#root {
+body {
+  margin: 0;
+  background-color: var(--bg-color);
+  background-image: var(--bg-image);
+  background-attachment: fixed;
   text-align: center;
   display: flex;
   flex-direction: column;
   height: 100%;
   min-height: 100vh;
+  color: var(--text-color);
+
+  --main-theme-color: #1d009e;
+  --bg-color: #090033;
+  --bg-image: url("../../public/dark-fon.jpg");
+  --transparent-color: rgb(9, 0, 51, 0.7);
+  --transparent-color-hover: rgba(119, 94, 229, 0.3);
+  --transparent-color-disabled: rgba(119, 94, 229, 0.2);
+  --text-color: #dedbff;
+  --accent-theme-color: #6e75ff;
+  --links-color-hover: #a0a5ff;
+  --main-color: #7378d8;
 }
 
-body {
-  margin: 0;
-  background-attachment: fixed;
+body[theme="light"] {
+  --main-theme-color: #ccc1ff;
+  --bg-color: #b4a6f3;
+  --bg-image: url("../../public/light-fon.jpg");
+  --transparent-color: rgba(180, 166, 243, 0.7);
+  --transparent-color-hover: rgba(233, 233, 233, 0.6);
+  --transparent-color-disabled: rgba(233, 233, 233, 0.5);
+  --text-color: #090033;
+  --accent-theme-color: #0009b1;
+  --links-color-hover: #4d54da;
+  --main-color: #7277ff;
 }
 
-.dark {
-  background-image: url("../../public/dark-fon.jpg");
-  background-color: var(--transparent-dark);
+.bg-color {
+  background-color: var(--transparent-color);
   background-blend-mode: darken;
   color: var(--text-light);
 }
 
-.light {
-  background-image: url("../../public/light-fon.jpg");
-  background-color: var(--transparent-light);
-  background-blend-mode: darken;
-  color: var(--text-dark);
-}
-
 a {
   font-weight: 500;
-  color: var(--links);
+  color: var(--main-color);
   text-decoration: inherit;
   transition: color 0.3s ease;
 }
@@ -88,7 +83,7 @@ button {
   font-size: 1em;
   font-weight: 400;
   font-family: inherit;
-  background-color: var(--main-buttons);
+  background-color: var(--main-color);
   cursor: pointer;
   transition: background-color 0.25s;
 }
@@ -103,7 +98,7 @@ p {
 }
 
 .error-message {
-  color: var(--accent-red);
+  color: var(--accent-color);
   font-weight: 300;
   position: absolute;
   top: 60%;
@@ -141,16 +136,16 @@ p {
   width: 40%;
 }
 
-.text-over-content {
+/* .text-over-content {
   position: absolute;
   top: 60%;
   left: 50%;
   transform: translate(-50%, -50%);
-  color: var(--text-light);
+  color: var(--text-color);
   font-size: 450px;
   padding: 0;
   margin: 0;
-}
+} */
 
 .turn-to-main-page-link {
   position: relative;
@@ -158,16 +153,189 @@ p {
   font-size: 20px;
 }
 
-@media (hover: hover) {
-  .dark a:hover {
-    color: var(--links-dark-hover);
-  }
+.toggle-theme__button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+}
 
-  .light a:hover {
-    color: var(--links-light-hover);
+.pagination {
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+}
+
+.pagination-numbers {
+  display: flex;
+  gap: 10px;
+}
+
+.pagination-link {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 5px;
+  border-radius: 5px;
+  font-size: 1em;
+  font-weight: 400;
+  font-family: inherit;
+  color: var(--main-theme-color);
+  border: 1px solid var(--main-theme-color);
+  background-color: var(--transparent-color);
+  cursor: pointer;
+  transition: background-color 0.4s ease;
+}
+
+.pagination-link_active {
+  background-color: var(--main-theme-color);
+  color: #000000;
+}
+
+.pagination-link_disabled {
+  pointer-events: none;
+  color: var(--transparent-color-disabled);
+  border: 1px solid var(--transparent-color-disabled);
+}
+
+.results {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  list-style: none;
+  margin: 20px auto;
+  gap: 20px;
+}
+
+.results__item {
+  text-align: start;
+  display: flex;
+  flex-direction: column;
+  border-radius: 10px;
+  box-shadow: 4px 4px 5px var(--border-shadow-color);
+  padding: 10px;
+  width: 260px;
+  height: 115px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.results__item {
+  background-color: var(--transparent-color);
+  border: 1px solid var(--main-theme-color);
+}
+
+.params__text {
+  display: flex;
+  flex-direction: column;
+}
+
+.name__text,
+.params__text {
+  font-weight: 500;
+}
+
+.name__item {
+  font-weight: 400;
+}
+
+.name__text {
+  color: var(--accent-theme-color);
+}
+
+.params__item {
+  color: var(--text-color);
+  font-weight: 300;
+}
+
+.search-form {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+}
+
+.search-input {
+  width: 500px;
+  height: 30px;
+}
+
+.search-button {
+  width: 100px;
+  height: 30px;
+}
+
+.search-section::after {
+  content: "";
+  display: block;
+  margin: 10px 0;
+}
+
+.search-section::after {
+  border-bottom: 1px solid var(--main-theme-color);
+}
+
+.results-section {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+  margin: 0 20px;
+  flex-grow: 1;
+}
+
+.result-section-with-detail {
+  display: flex;
+  gap: 10px;
+  flex-grow: 1;
+}
+
+.details {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  margin: 5px;
+}
+
+.flyout {
+  position: fixed;
+  bottom: 200px;
+  width: 300px;
+
+  border-radius: 8px;
+  padding: 10px;
+  z-index: 1;
+}
+
+.flyout {
+  background-color: var(--bg-color);
+  border: 1px solid var(--main-theme-color);
+}
+
+.flyout__buttons {
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+  padding: 10px 0;
+}
+
+@media (hover: hover) {
+  a:hover {
+    color: var(--links-color-hover);
   }
 
   button:hover {
-    background-color: var(--main-buttons-hover);
+    background-color: var(--main-color-hover);
+  }
+
+  .pagination-link:hover {
+    background-color: var(--main-theme-color);
+    color: #000000;
+  }
+
+  .results__item:hover {
+    background-color: var(--transparent-color-hover);
+    box-shadow: 4px 4px 20px var(--border-shadow-color);
   }
 }
